add hybrid search to topic channel pages (#3401)
* move hybrid search to its own component

* add vector search to topic pages

* fix test for vector endpoint mock

* fix mock for vector endpoint

* adding test for topic pages

* fix test

* Potential fix for pull request finding


* Potential fix for pull request finding


* fix format

---------
