+++ frontends/main/src/app-pages/ChannelPage/ChannelSearch.tsx
@@ -5,6 +5,7 @@ import { useResourceSearchParams } from "@mitodl/course-search-utils"
 import type { Facets, BooleanFacets } from "@mitodl/course-search-utils"
 import { useSearchParams } from "@mitodl/course-search-utils/next"
 import SearchDisplay from "@/page-components/SearchDisplay/SearchDisplay"
+import HybridSearchDisplay from "@/page-components/SearchDisplay/HybridSearchDisplay"
 import { Container, styled } from "ol-components"
 import { VisuallyHidden } from "@mitodl/smoot-design"
 import { SearchField } from "@/page-components/SearchField/SearchField"
@@ -99,7 +100,8 @@ const ChannelSearch: React.FC<ChannelSearchProps> = ({
   useEffect(() => {
     setCurrentText(params.q ?? "")
   }, [params, setCurrentText])
-
+  const ChannelSearchDisplay =
+    channelType === ChannelTypeEnum.Topic ? HybridSearchDisplay : SearchDisplay
   return (
     <section>
       <VisuallyHidden as="h2">Search within {channelTitle}</VisuallyHidden>
@@ -118,7 +120,7 @@ const ChannelSearch: React.FC<ChannelSearchProps> = ({
         />
       </SearchInputContainer>
 
-      <SearchDisplay
+      <ChannelSearchDisplay
         resultsHeadingEl="h3"
         filterHeadingEl="h3"
         page={page}
